move PrivateKey methods to Unredacted impl block in-place
