+++ src/unredacted.rs
@@ -1,103 +1,15 @@
-//! Unredacted wrapper for [`PrivateKey`].
+//! Unredacted wrapper used to gate `Display` and `Serialize` on private-key
+//! material.
 //!
-//! `PrivateKey` does not implement [`Display`] or `Serialize` directly. To
-//! render the encoded strkey form or serialize via `serde`, callers wrap
-//! the value in [`Unredacted`].
-//!
-//! `Debug` on the bare type emits the redacted form, and `Deserialize` is
-//! implemented (input only — parsing a strkey string does not leak), so
-//! neither requires a wrapper.
+//! [`PrivateKey`](crate::ed25519::PrivateKey) does not implement
+//! [`Display`](core::fmt::Display) or `Serialize` directly. Callers wrap the
+//! value in [`Unredacted`] to render the encoded strkey form or serialize
+//! via `serde`.
 
-use core::fmt::{self, Debug, Display, Formatter};
-use core::str::FromStr;
-
-use heapless::String as HeaplessString;
-use zeroize::Zeroizing;
-
-use crate::{convert::encode_zeroizing, ed25519::PrivateKey, error::DecodeError, version};
-
-/// Wraps a [`PrivateKey`] so it can be rendered or serialized in its full
-/// strkey string form.
+/// Wrapper that opts a value in to formatting or serialization that would
+/// otherwise expose private-key bytes.
 #[derive(Clone, Copy, Hash, PartialEq, Eq, PartialOrd, Ord)]
 pub struct Unredacted<T>(pub T);
-
-impl Unredacted<&PrivateKey> {
-    /// Encodes this private key to its strkey string form.
-    ///
-    /// # Zeroize
-    ///
-    /// The intermediate scratch buffers used during encoding are zeroed on
-    /// drop, but the returned `String` itself is plain — its bytes are not
-    /// zeroed when the value is dropped. Use
-    /// [`write_string`](Self::write_string) for zeroizing.
-    pub fn to_string(&self) -> HeaplessString<{ PrivateKey::ENCODED_LEN }> {
-        let mut zeroizing: Zeroizing<HeaplessString<{ PrivateKey::ENCODED_LEN }>> =
-            Zeroizing::new(HeaplessString::new());
-        self.write_string(&mut zeroizing);
-        let mut out: HeaplessString<{ PrivateKey::ENCODED_LEN }> = HeaplessString::new();
-        out.push_str(&zeroizing).unwrap();
-        out
-    }
-
-    /// Encodes this private key to its strkey string form, writing the
-    /// result into the caller-provided buffer.
-    ///
-    /// # Zeroize
-    ///
-    /// The intermediate scratch buffers used during encoding are wrapped in
-    /// [`Zeroizing`] and zeroed on drop, and the encoded bytes are written
-    /// directly into `out` rather than returned by value, so no copy is left
-    /// on this method's stack frame.
-    pub fn write_string(&self, out: &mut Zeroizing<HeaplessString<{ PrivateKey::ENCODED_LEN }>>) {
-        encode_zeroizing::<
-            { PrivateKey::PAYLOAD_LEN },
-            { PrivateKey::BINARY_LEN },
-            { PrivateKey::ENCODED_LEN },
-        >(version::PRIVATE_KEY_ED25519, &self.0 .0, out);
-    }
-}
-
-impl Display for Unredacted<&PrivateKey> {
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        let mut buf: Zeroizing<HeaplessString<{ PrivateKey::ENCODED_LEN }>> =
-            Zeroizing::new(HeaplessString::new());
-        self.write_string(&mut buf);
-        f.write_str(&buf)
-    }
-}
-
-impl Debug for Unredacted<&PrivateKey> {
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        f.write_str("PrivateKey(")?;
-        for b in &self.0 .0 {
-            write!(f, "{b:02x}")?;
-        }
-        f.write_str(")")
-    }
-}
-
-// --- Owned Display/Debug (delegate to borrowed form) ---
-
-impl Display for Unredacted<PrivateKey> {
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        Display::fmt(&Unredacted(&self.0), f)
-    }
-}
-
-impl Debug for Unredacted<PrivateKey> {
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        Debug::fmt(&Unredacted(&self.0), f)
-    }
-}
-
-// --- FromStr ---
-
-impl FromStr for Unredacted<PrivateKey> {
-    type Err = DecodeError;
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        PrivateKey::from_string(s).map(Unredacted)
-    }
-}
 
 #[cfg(feature = "serde")]
 mod serde_impl {
